+++ src/corecel/sys/Counter.hh
@@ -13,20 +13,13 @@
 
 namespace celeritas
 {
-//---------------------------------------------------------------------------//
 #if CELERITAS_USE_PERFETTO
-/*!
- * Simple tracing counter.
- * Records a named value at the current timestamp which
- * can then be displayed on a timeline. Only supported on host, this compiles
- * but is a noop on device.
- * See https://perfetto.dev/docs/instrumentation/track-events#counters
- * \tparam T Arithmetic counter type
- */
+//---------------------------------------------------------------------------//
+// Simple tracing counter
 template<class T>
 CELER_FUNCTION void trace_counter(char const* name, T value);
 #else
-// noop
+//! No tracing backend - noop
 template<class T>
 CELER_FUNCTION inline void trace_counter(char const*, T)
 {
