+++ src/corecel/sys/TracingSession.cc
@@ -3,7 +3,7 @@
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file corecel/sys/TracingSession.perfetto.cc
+//! \file corecel/sys/TracingSession.cc
 //! \brief RAII class for managing a perfetto session and its resources.
 //---------------------------------------------------------------------------//
 #include "TracingSession.hh"
@@ -23,6 +23,7 @@ namespace
 //---------------------------------------------------------------------------//
 using celeritas::TracingMode;
 
+//---------------------------------------------------------------------------//
 /*!
  * Initialize the session for the given mode if profiling is enabled.
  */
@@ -49,6 +50,7 @@ std::unique_ptr<perfetto::TracingSession> initialize_session(TracingMode mode)
     return perfetto::Tracing::NewTrace();
 }
 
+//---------------------------------------------------------------------------//
 /*!
  * Configure the session to record Celeritas track events.
  */
